Refactor theme, navigation, and homepage image
Updated the site theme from 'caramellatte' to 'fantasy' in HTML and CSS. Refactored tag styling in .eleventy.js and src/pages/.eleventy.js for consistency. Removed the /apps.html page and its navigation link, and moved old blog posts to an 'old' directory. Enhanced the homepage image with a hover effect to show a new photo. Adjusted footer and header layouts for improved alignment and clarity.